requirements.txt
# General
numpy
python-dotenv

# Groq client
groq

# LangChain ecosystem (community split)
langchain-community>=0.0.30
langchain-openai>=0.0.5
langgraph>=0.0.23

# Optional
openai
langchain_groq